execution/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
